QABot/pinecone_django/requirements.txt
django
sentence-transformers
pinecone-client
openai
